Go back to declaring a function and returning its name :(
Really sorry to create all this unnecessary work for you, but after
working with this code in more depth, I've become convinced the only
viable approach is the original approach, because IE8 simply can't handle
named function expressions at all.

The final, unavoidable gotcha was that, given this function definition:

  var foo = (function () {
    return function foo () {
      return foo;
    };
  })();

the following assertion fails, if you can believe it:

  assert.strictEqual(foo(), foo);

The reason is that the foo function object that leaks into the IIFE scope
(and thus the foo visible in the body of the innermost function) is not
even the same object as the function object returned from the IIFE!

In particular, this causes babelHelpers.classCallCheck to fail during
class instantiation, since this.constructor differs from the constructor
function invoked by new.